Implemented a hack to detect failed Imgur extension guesses (e.g. gallery link which gets a .png extension but is actually a GIFV)
Imgur's "removed" image is now always loaded, and its dimensions are compared onload with those of any Imgur image for which we guessed the extension.

Displaying link text as the replacement image's title now to make it aesier to roll back to the original link contents if a missing image is detected.